Moved page naming convention to be that of the component it exports
