+++ frontend/src/styles/Dashboard.css
