+++ frontend/src/styles/Home.css
